Add components folder and moved appjs to it
